+++ src/components/ui/index.tsx
@@ -1,0 +1,5 @@
+export * from "./button";
+export * from "./dropdown-menu";
+export * from "./form";
+export * from "./label";
+export * from "./input";