+++ CHANGELOG.md
@@ -9,5 +9,6 @@ NOTE: all lowercase and no ending punctuation marks
 
 ## Next release v1.0
 
+- [#3](https://github.com/cyril-deguzman/ntzwlt/issues/3) app(pages): add home and login pages with dark mode
 - [#2](https://github.com/cyril-deguzman/ntzwlt/issues/2) app(dependency): add zod, react query, and shadcn
 - [#1](https://github.com/cyril-deguzman/ntzwlt/issues/1) repo(markdown): modify readme and add changelog